+++ lib/Sema/CMakeLists.txt
@@ -5,6 +5,7 @@ endif()
 
 add_swift_library(swiftSema STATIC
   CSApply.cpp
+  CSBindings.cpp
   CSDiag.cpp
   CSGen.cpp
   CSPropagate.cpp
